Added flush method to impressions and track manager. Fixed name for impressions manager. Created TrackManager protocol
